requirements.txt
torch>=1.12.0
torchvision>=0.13.0
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.2
nibabel>=3.2.0
pyradiomics>=3.0
matplotlib>=3.4.0
opencv-python>=4.5.3
tqdm>=4.62.0
Pillow>=8.2.0
SimpleITK>=2.1.0
xgboost>=1.6.0
